+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Expanded punctuation to support parenthesis in both Parmaitë and Annatar
   including apostraphe, comma, colon, and full-stop. 
 - Added OE and IE diphthongs in the mode for general use.
+- Adds all three representations of Y: for Quenya, Sindarin, and English.
+### Fixed
+- Fixes the representation of J, CK, EE, and Y for English in the mode for general use.
+- Fixes the erroneous use of wilya for the silent carrier anna, when A appears alone.
 - Changed the default representation for the PH cluster in the mode for general
   use to retain a hint of the unusual orthography for formen by using the
   combined formen-parma tengwa as is attested.